convert cod3.c to D

+++ src/dmd/backend/dwarf.c
@@ -366,7 +366,7 @@ static Outbuffer cfa_buf;               // CFA instructions
  * Params:
  *      location = offset from the start of the function
  */
-void dwarf_CFA_set_loc(size_t location)
+void dwarf_CFA_set_loc(unsigned location)
 {
     assert(location >= CFA_state_current.location);
     size_t inc = location - CFA_state_current.location;
@@ -2991,7 +2991,7 @@ void dwarf_except_gentables(Funcsym *sfunc, unsigned startoffset, unsigned retof
 }
 
 #else
-void dwarf_CFA_set_loc(size_t location) { }
+void dwarf_CFA_set_loc(unsigned location) { }
 void dwarf_CFA_set_reg_offset(int reg, int offset) { }
 void dwarf_CFA_offset(int reg, int offset) { }
 void dwarf_except_gentables(Funcsym *sfunc, unsigned startoffset, unsigned retoffset) { }
